вынос find_sw_vendor
