+++ func/check_ip_address_function.py
@@ -1,0 +1,9 @@
+import re
+
+def check_ip_address(ip_address):
+    ip_pattern = re.compile(r'^(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?).(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?).(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?).(25[0-5]|2[0-4][0-9]|[01]?[0-9][0-9]?)$')  # Regular expression for checking the IP address
+
+    if ip_pattern.match(ip_address):
+        return True
+    else:
+        return False
